IssueNo:https://gitcode.com/openharmony/kernel_liteos_a/issues/1081 Feature or Bugfix: Bugfix ~ Binary Source: No Signed-off-by: yang-pangyuan <yangpangyuan1@h-partners.com>

+++ testsuites/kernel/sample/kernel_base/core/task/smp/It_smp_los_task_068.c
@@ -55,7 +55,7 @@ static void HwiF01(void)
 static void TaskF01(void)
 {
     UINT32 ret;
-    UINT32 readbuf[50];
+    UINT8 readbuf[50];
 
     LOS_AtomicInc(&g_testCount);
